% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languagelayeR.R
\name{detect_language}
\alias{detect_language}
\title{Detect Language of the given String}
\usage{
detect_language(access_key, query, data_frame = TRUE)
}
\arguments{
\item{access_key}{LanguageLayer API Access Key}

\item{query}{Character String whose language has to be detected}

\item{data_frame}{List is returned if FALSE dataframe if TRUE (default TRUE)}
}
\value{
A list or dataframe with the detected language and other parameters
}
\description{
Detect Language of the given String
}
\examples{
get_valid_currencies()
}
