% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_supported_lang.R
\name{get_supported_lang}
\alias{get_supported_lang}
\title{Get supported languages}
\usage{
get_supported_lang(api_key)
}
\arguments{
\item{api_key}{Your API key.}
}
\value{
Returns a data.frame with language_code and language_name.
}
\description{
Get all current available languages on the languagelayer API.
}
\details{
Returns a data.frame with the available languages.
}
\note{
Before running a function of this package for the first time, you need to set your API key using the \code{\link{setApiKey}} function.
}
\examples{
get_supported_lang(api_key = "yourapikey")
}
