% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSupportedLanguage.R
\name{getSupportedLanguage}
\alias{getSupportedLanguage}
\title{Get supported languages}
\usage{
getSupportedLanguage()
}
\value{
Returns a data.frame with language_code and language_name.
}
\description{
Get all current available languages on the languagelayer API.
}
\details{
Returns a data.frame with the available languages.
}
\note{
Before running a function of this package for the first time, you need to set your API key using the \code{\link{setApiKey}} function.
}
\examples{
getSupportedLanguage()
}

