% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languagelayeR-package.r
\docType{package}
\name{languagelayeR}
\alias{languagelayeR}
\alias{languagelayeR-package}
\title{languagelayeR: Detect Language Of The Given Text Using 'LanguageLayer API'}
\description{
Detect Language type of the given text using LanguageLayer API \url{https://languagelayer.com}.
The languagelayer API relies on an ever-advancing, powerful and AI-based
(Artificial Intelligence) detection algorithm.
}
\section{See Also}{

Useful links:
\itemize{
\item \href{https://languagelayer.com}{LanguageLayer API}
\item \url{https://github.com/amrrs/languagelayeR}
\item Report bugs at \url{https://github.com/amrrs/languagelayeR/issues}
}
}

