% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setApiKey.R
\name{setApiKey}
\alias{setApiKey}
\title{Set API Key}
\usage{
setApiKey(apikey)
}
\arguments{
\item{apikey}{You can find your API key on your dashboard at \url{https://languagelayer.com/dashboard}.}
}
\description{
Set your API key.
}
\details{
Takes a character string containing your API key.
}
