% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setApiKey.R
\name{setApiKey}
\alias{setApiKey}
\title{Set API Key}
\usage{
setApiKey(apikey)
}
\arguments{
\item{apikey}{the API key you can get from your languagelayer dashboard at \url{https://languagelayer.com/dashboard}.}
}
\description{
Set your API key to access the languagelayer API
}
\details{
Takes a character string containing your API key.
}
\examples{
setApiKey(apikey = "yourapikey")
}

